Add metric prefixes
Some graphite services require that the API key for the
service (e.g. hostedgraphite) is inserted into the metric name
itself. This patch allows to insert this into the metric name.

Otherwise, it is required that some kind of proxy/relay is used in order
to accomplish this (such as graphite-carbon-relay).